query: big rework of the paging model
Instead of sending data from C++ to javascript as fast as possible, only
send data a page at a time and pass a reference to the query object so the
client can re-issue the query for another page of data.

This involved refactoring the implmentation so that Query is a proper
wrapped object that can be referenced from Javascript.