async-future: go back to the event loop after each wakeup
To avoid potentially starving other I/O when the process is busy
handling async future callbacks, make a snapshot of the pending
queue when async_ready is called before running through the queue
to issue the callbacks.

If additional events arrive while processing the first batch, then
uv_async_send will also be called which will trigger it to wake up
another time to process the rest of the events.