types: add optional encoding support for parameter and result types
In addition to allowing the caller to override the cassandra type for
a parameter, support additional encoding flags. Use this to make the
support for passing a bigint as a low/high pair optional instead of
mandatory.

Also add analogous support for passing an array of result_types as an
option to query so that the caller can control how the results are
encoded.